requirements.txt
moviepy==1.0.3
requests
python-dotenv
fastapi
uvicorn
python-multipart
pyautogui
openai>=1.0.0
